+++ js/utils.js
@@ -1,0 +1,28 @@
+function getRandomInteger (a, b) {
+  const lower = Math.ceil(Math.min(Math.abs(a), Math.abs(b)));
+  const upper = Math.floor(Math.max(Math.abs(a), Math.abs(b)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+
+function getRandomArrayElement (elements) {
+  return elements[getRandomInteger(0, elements.length - 1)];
+}
+
+function createRandomId (a, b) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(a, b);
+    if (previousValues.length >= (b - a + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(a, b);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export {getRandomInteger, getRandomArrayElement, createRandomId};
